requirements.txt
# Data Collection
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.15.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Text Processing & NLP
nltk==3.8.1

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.3

# Dashboard
streamlit==1.29.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1